requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
opencv-python>=4.6.0
numpy>=1.23.0

# Web Framework
flask>=2.3.0
werkzeug>=2.3.0

# CLI
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
